requirements.txt
torch
torchvision
gradio
numpy
tqdm
requests
ninja
click
pyspng
scipy
pillow
